Refactored `stream_attack` to remove scattered `From` impls
