+++ leech/src/rpc/mod.rs
@@ -18,23 +18,17 @@ use crate::rpc::attacks::Attacks;
 use crate::rpc::rpc_attacks::req_attack_service_server::ReqAttackServiceServer;
 
 pub mod attacks;
-mod utils;
 
 /// Missing docs are allowed, as the code gets auto generated by tonic
 #[allow(missing_docs)]
 pub mod rpc_attacks {
-    use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
+    use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
     use std::ops::RangeInclusive;
 
     use ipnetwork::{IpNetwork, Ipv4Network, Ipv6Network};
     use tonic::Status;
 
-    use crate::modules::bruteforce_subdomains::BruteforceSubdomainResult;
-    use crate::modules::dns::DnsRecordResult;
-    use crate::rpc::rpc_attacks::shared::dns_record::Record;
-    use crate::rpc::rpc_attacks::shared::{
-        Aaaa, Address, DnsRecord, GenericRecord, Ipv4, Ipv6, Net, NetOrAddress, A,
-    };
+    use crate::rpc::rpc_attacks::shared::{Address, Ipv4, Ipv6, Net, NetOrAddress};
 
     pub mod shared {
         tonic::include_proto!("attacks.shared");
@@ -112,92 +106,6 @@ pub mod rpc_attacks {
                     }
                 }
                 shared::net_or_address::NetOrAddress::Net(x) => x.into(),
-            }
-        }
-    }
-
-    impl From<BruteforceSubdomainResult> for BruteforceSubdomainResponse {
-        fn from(value: BruteforceSubdomainResult) -> Self {
-            Self {
-                record: Some(match value {
-                    BruteforceSubdomainResult::A { source, target } => DnsRecord {
-                        record: Some(Record::A(A {
-                            source,
-                            to: Some(target.into()),
-                        })),
-                    },
-                    BruteforceSubdomainResult::Aaaa { source, target } => DnsRecord {
-                        record: Some(Record::Aaaa(Aaaa {
-                            source,
-                            to: Some(target.into()),
-                        })),
-                    },
-                    BruteforceSubdomainResult::Cname { source, target } => DnsRecord {
-                        record: Some(Record::Cname(GenericRecord { source, to: target })),
-                    },
-                }),
-            }
-        }
-    }
-
-    impl From<DnsRecordResult> for DnsResolutionResponse {
-        fn from(value: DnsRecordResult) -> Self {
-            Self {
-                record: Some(match value {
-                    DnsRecordResult::A { source, target } => DnsRecord {
-                        record: Some(Record::A(A {
-                            source,
-                            to: Some(target.into()),
-                        })),
-                    },
-                    DnsRecordResult::Aaaa { source, target } => DnsRecord {
-                        record: Some(Record::Aaaa(Aaaa {
-                            source,
-                            to: Some(target.into()),
-                        })),
-                    },
-                    DnsRecordResult::CAA { source, target } => DnsRecord {
-                        record: Some(Record::Caa(GenericRecord { source, to: target })),
-                    },
-                    DnsRecordResult::Cname { source, target } => DnsRecord {
-                        record: Some(Record::Cname(GenericRecord { source, to: target })),
-                    },
-                    DnsRecordResult::Mx { source, target } => DnsRecord {
-                        record: Some(Record::Mx(GenericRecord { source, to: target })),
-                    },
-                    DnsRecordResult::Tlsa { source, target } => DnsRecord {
-                        record: Some(Record::Tlsa(GenericRecord { source, to: target })),
-                    },
-                    DnsRecordResult::Txt { source, target } => DnsRecord {
-                        record: Some(Record::Txt(GenericRecord { source, to: target })),
-                    },
-                }),
-            }
-        }
-    }
-
-    impl From<SocketAddr> for TcpPortScanResponse {
-        fn from(value: SocketAddr) -> Self {
-            let address = match value {
-                SocketAddr::V4(v) => Address {
-                    address: Some(shared::address::Address::Ipv4((*v.ip()).into())),
-                },
-                SocketAddr::V6(v) => Address {
-                    address: Some(shared::address::Address::Ipv6((*v.ip()).into())),
-                },
-            };
-
-            Self {
-                address: Some(address),
-                port: value.port() as u32,
-            }
-        }
-    }
-
-    impl From<IpAddr> for HostsAliveResponse {
-        fn from(value: IpAddr) -> Self {
-            Self {
-                host: Some(value.into()),
             }
         }
     }
